Refactor updater out of MenuBar and into UpdateManager
