+++ Cairo Desktop/Cairo Desktop/Cairo Desktop.csproj
@@ -120,6 +120,7 @@
     <Compile Include="SupportingClasses\FullScreenApp.cs" />
     <Compile Include="SupportingClasses\PicSumWallpaperClient.cs" />
     <Compile Include="SupportingClasses\ScreenSetupReason.cs" />
+    <Compile Include="SupportingClasses\UpdateManager.cs" />
     <Compile Include="SupportingClasses\WindowManager.cs" />
     <Compile Include="SupportingClasses\FilePathValidationRule.cs" />
     <Compile Include="SupportingClasses\FolderHelper.cs" />
