+++ Cairo Desktop/Cairo Desktop/SupportingClasses/UpdateManager.cs
@@ -1,0 +1,40 @@
+﻿using CairoDesktop.Common.DesignPatterns;
+using CairoDesktop.Interop;
+
+namespace CairoDesktop.SupportingClasses
+{
+    public class UpdateManager : SingletonObject<UpdateManager>
+    {
+        private const string UpdateUrl = "https://cairoshell.github.io/appdescriptor.rss";
+
+        private bool isInitialized;
+        private WinSparkle.win_sparkle_can_shutdown_callback_t canShutdownDelegate;
+        private WinSparkle.win_sparkle_shutdown_request_callback_t shutdownDelegate;
+
+        private UpdateManager() { }
+
+        public void Initialize()
+        {
+            if (!isInitialized)
+            {
+                WinSparkle.win_sparkle_set_appcast_url(UpdateUrl);
+                canShutdownDelegate = canShutdown;
+                shutdownDelegate = Startup.Shutdown;
+                WinSparkle.win_sparkle_set_can_shutdown_callback(canShutdownDelegate);
+                WinSparkle.win_sparkle_set_shutdown_request_callback(shutdownDelegate);
+                WinSparkle.win_sparkle_init();
+                isInitialized = true;
+            }
+        }
+
+        public void CheckForUpdates()
+        {
+            WinSparkle.win_sparkle_check_update_with_ui();
+        }
+
+        private int canShutdown()
+        {
+            return 1;
+        }
+    }
+}
